+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes to Calva.
 ## [Unreleased]
 
 - [Make it possible to select several output destinations](https://github.com/BetterThanTomorrow/calva/issues/2819)
+- [Support file paths as output destinations](https://github.com/BetterThanTomorrow/calva/issues/3185)
 
 ## [2.0.576] - 2026-04-25